still working on validation
There was no real bug with the last push. I just had it set to harware trig when actually I was software triggering. And in all the fuss I put the device in state with an instruction that didn't exist, so it hung. I just called reset_output_coordinator. All good.